requirements.txt
requests==2.31.0
pyyaml==6.0.1
pendulum==3.0.0
pytest==8.0.0
pytest-mock==3.12.0